[minor] remove Dom/Interactive extensions
